translator: hoist divisions to correct Rust truncated div/mod (#37)
Cora's div/mod are Euclidean but Rust's truncate toward zero, so they
diverge for negative dividends. Hoist each division to its own program
point that binds a fresh variable to the sign-corrected value, instead of
substituting the correction inline (which explodes at nested divisions).

- width literal-only divisions from their context type, else i32
- while: re-run the hoisted condition each iteration via the back-edge
- name hoist vars `$div` (`$` is legal in Cora, illegal in Rust, so it can't
  clash with a source variable); verified Cora parses and proves them
- regenerate division goldens; add while_div

+++ src/test/resources/benchmarks/count_up.rs
@@ -1,6 +1,6 @@
 // cora: MAYBE
-fn count_up(x: i32) -> i32 {
-    let mut y: i32 = 0;
+fn count_up(x: i16) -> i16 {
+    let mut y: i16 = 0;
     while x > y {
         y = y + 1;
     }
